Splitting up the tags and filters for templates.
Helps to keep things separated and organized.
